src/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
python-dotenv
pymongo